+++ src/egkeys.c
@@ -313,7 +313,7 @@ void selectMissile() {
 }
 
 // ==== seg000:0xda35 ====
-void makeSound(int soundId, int priority) {
+void makeSound(int16 soundId, int16 priority) {
     if (priority >= g_axisInputAccum[2]) {
         if (g_ejectState == 0 || priority > 1) {
             audio_playSound(soundId);
@@ -323,7 +323,7 @@ void makeSound(int soundId, int priority) {
 }
 
 // ==== seg000:0xda5f ====
-void playVoiceCue(int weaponIdx) {
+void playVoiceCue(int16 weaponIdx) {
     if (g_axisInputAccum[2] < 2 && g_ejectState == 0 &&
         (unsigned)voiceCueThresholds[weaponIdx] < (unsigned)f15DgtlResult) {
         audio_playSample(weaponIdx);
@@ -354,7 +354,7 @@ void recalcTimeScale(void) {
 
 // ==== seg000:0xdb2b ====
 void setupLodDistances(void) {
-    int lod;
+    int16 lod;
     /* Detail level 4 ("extended"): every object keeps its finest model and is
        never distance-culled, so the long-range terrain still resolves fully.
        Levels 0-3 use the original per-LOD distance thresholds. */
