+++ src/egkeys.h
@@ -52,8 +52,8 @@ enum ScanCode : uint16 {
 };
 
 void keyDispatch(uint16 scanCode);
-void makeSound(int soundId, int priority);
-void playVoiceCue(int weaponIdx);
+void makeSound(int16 soundId, int16 priority);
+void playVoiceCue(int16 weaponIdx);
 void updateEngineSound(void);
 void recalcTimeScale(void);
 void setupLodDistances(void);
